pyproject.toml
[project]
name = "perflaunch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pynput>=1.8.1",
    "pystray>=0.19.5",
]
